periodbase: refactored to use submodules, tested and works fine
