+++ astrobase/periodbase/spdm.py
@@ -1,0 +1,373 @@
+#!/usr/bin/env python
+
+'''spdm.py - Waqas Bhatti (wbhatti@astro.princeton.edu) - Jan 2017
+
+Contains the Stellingwerf (1978) phase-dispersion minimization period-search
+algorithm implementation for periodbase.
+
+'''
+
+
+from multiprocessing import Pool, cpu_count
+import logging
+from datetime import datetime
+from traceback import format_exc
+
+import numpy as np
+
+# import these to avoid lookup overhead
+from numpy import nan as npnan, sum as npsum, abs as npabs, \
+    roll as nproll, isfinite as npisfinite, std as npstd, \
+    sign as npsign, sqrt as npsqrt, median as npmedian, \
+    array as nparray, percentile as nppercentile, \
+    polyfit as nppolyfit, var as npvar, max as npmax, min as npmin, \
+    log10 as nplog10, arange as nparange, pi as MPI, floor as npfloor, \
+    argsort as npargsort, cos as npcos, sin as npsin, tan as nptan, \
+    where as npwhere, linspace as nplinspace, \
+    zeros_like as npzeros_like, full_like as npfull_like, \
+    arctan as nparctan, nanargmax as npnanargmax, nanargmin as npnanargmin, \
+    empty as npempty, ceil as npceil, mean as npmean, \
+    digitize as npdigitize, unique as npunique, \
+    argmax as npargmax, argmin as npargmin
+
+
+#############
+## LOGGING ##
+#############
+
+# setup a logger
+LOGGER = None
+
+def set_logger_parent(parent_name):
+    globals()['LOGGER'] = logging.getLogger('%s.spdm' % parent_name)
+
+def LOGDEBUG(message):
+    if LOGGER:
+        LOGGER.debug(message)
+    elif DEBUG:
+        print('%sZ [DBUG]: %s' % (datetime.utcnow().isoformat(), message))
+
+def LOGINFO(message):
+    if LOGGER:
+        LOGGER.info(message)
+    else:
+        print('%sZ [INFO]: %s' % (datetime.utcnow().isoformat(), message))
+
+def LOGERROR(message):
+    if LOGGER:
+        LOGGER.error(message)
+    else:
+        print('%sZ [ERR!]: %s' % (datetime.utcnow().isoformat(), message))
+
+def LOGWARNING(message):
+    if LOGGER:
+        LOGGER.warning(message)
+    else:
+        print('%sZ [WRN!]: %s' % (datetime.utcnow().isoformat(), message))
+
+def LOGEXCEPTION(message):
+    if LOGGER:
+        LOGGER.exception(message)
+    else:
+        print(
+            '%sZ [EXC!]: %s\nexception was: %s' % (
+                datetime.utcnow().isoformat(),
+                message, format_exc()
+                )
+            )
+
+
+###################
+## LOCAL IMPORTS ##
+###################
+
+from ..lcmath import phase_magseries, sigclip_magseries, time_bin_magseries, \
+    phase_bin_magseries
+
+from . import get_frequency_grid
+
+
+############
+## CONFIG ##
+############
+
+NCPUS = cpu_count()
+
+
+####################################################################
+## PHASE DISPERSION MINIMIZATION (Stellingwerf+ 1978, 2011, 2013) ##
+####################################################################
+
+def stellingwerf_pdm_theta(times, mags, errs, frequency,
+                           binsize=0.05, minbin=9):
+    '''
+    This calculates the Stellingwerf PDM theta value at a test frequency.
+
+    '''
+
+    period = 1.0/frequency
+    fold_time = times[0]
+
+    phased = phase_magseries(times,
+                             mags,
+                             period,
+                             fold_time,
+                             wrap=False,
+                             sort=True)
+
+    phases = phased['phase']
+    pmags = phased['mags']
+    bins = np.arange(0.0, 1.0, binsize)
+    nbins = bins.size
+
+    binnedphaseinds = npdigitize(phases, bins)
+
+    binvariances = []
+    binndets = []
+    goodbins = 0
+
+    for x in npunique(binnedphaseinds):
+
+        thisbin_inds = binnedphaseinds == x
+        thisbin_phases = phases[thisbin_inds]
+        thisbin_mags = pmags[thisbin_inds]
+
+        if thisbin_mags.size > minbin:
+            thisbin_variance = npvar(thisbin_mags,ddof=1)
+            binvariances.append(thisbin_variance)
+            binndets.append(thisbin_mags.size)
+            goodbins = goodbins + 1
+
+    # now calculate theta
+    binvariances = nparray(binvariances)
+    binndets = nparray(binndets)
+
+    theta_top = npsum(binvariances*(binndets - 1)) / (npsum(binndets) -
+                                                      goodbins)
+    theta_bot = npvar(pmags,ddof=1)
+    theta = theta_top/theta_bot
+
+    return theta
+
+
+
+def stellingwerf_pdm_worker(task):
+    '''
+    This is a parallel worker for the function below.
+
+    task[0] = times
+    task[1] = mags
+    task[2] = errs
+    task[3] = frequency
+    task[4] = binsize
+    task[5] = minbin
+
+    '''
+
+    times, mags, errs, frequency, binsize, minbin = task
+
+    try:
+
+        theta = stellingwerf_pdm_theta(times, mags, errs, frequency,
+                                       binsize=binsize, minbin=minbin)
+
+        return theta
+
+    except Exception as e:
+
+        return npnan
+
+
+
+def stellingwerf_pdm(times,
+                     mags,
+                     errs,
+                     autofreq=True,
+                     startp=None,
+                     endp=None,
+                     normalize=False,
+                     stepsize=1.0e-4,
+                     phasebinsize=0.05,
+                     mindetperbin=9,
+                     nbestpeaks=5,
+                     periodepsilon=0.1, # 0.1
+                     sigclip=10.0,
+                     nworkers=None):
+    '''This runs a parallel Stellingwerf PDM period search.
+
+    '''
+
+    # get rid of nans first
+    find = np.isfinite(times) & np.isfinite(mags) & np.isfinite(errs)
+    ftimes = times[find]
+    fmags = mags[find]
+    ferrs = errs[find]
+
+    if len(ftimes) > 9 and len(fmags) > 9 and len(ferrs) > 9:
+
+        # get the median and stdev = 1.483 x MAD
+        median_mag = np.median(fmags)
+        stddev_mag = (np.median(np.abs(fmags - median_mag))) * 1.483
+
+        # sigclip next
+        if sigclip:
+
+            sigind = (np.abs(fmags - median_mag)) < (sigclip * stddev_mag)
+
+            stimes = ftimes[sigind]
+            smags = fmags[sigind]
+            serrs = ferrs[sigind]
+
+            LOGINFO('sigclip = %s: before = %s observations, '
+                    'after = %s observations' %
+                    (sigclip, len(times), len(stimes)))
+
+        else:
+
+            stimes = ftimes
+            smags = fmags
+            serrs = ferrs
+
+        # make sure there are enough points to calculate a spectrum
+        if len(stimes) > 9 and len(smags) > 9 and len(serrs) > 9:
+
+            # get the frequencies to use
+            if startp:
+                endf = 1.0/startp
+            else:
+                # default start period is 0.1 day
+                endf = 1.0/0.1
+
+            if endp:
+                startf = 1.0/endp
+            else:
+                # default end period is length of time series divided by 2
+                startf = 1.0/(stimes.max() - stimes.min())
+
+            # if we're not using autofreq, then use the provided frequencies
+            if not autofreq:
+                frequencies = np.arange(startf, endf, stepsize)
+                LOGINFO(
+                    'using %s frequency points, start P = %.3f, end P = %.3f' %
+                    (frequencies.size, 1.0/endf, 1.0/startf)
+                )
+            else:
+                # this gets an automatic grid of frequencies to use
+                frequencies = get_frequency_grid(stimes,
+                                                 minfreq=startf,
+                                                 maxfreq=endf)
+                LOGINFO(
+                    'using autofreq with %s frequency points, '
+                    'start P = %.3f, end P = %.3f' %
+                    (frequencies.size,
+                     1.0/frequencies.max(),
+                     1.0/frequencies.min())
+                )
+
+            # map to parallel workers
+            if (not nworkers) or (nworkers > NCPUS):
+                nworkers = NCPUS
+                LOGINFO('using %s workers...' % nworkers)
+
+            pool = Pool(nworkers)
+
+            # renormalize the working mags to zero and scale them so that the
+            # variance = 1 for use with our LSP functions
+            if normalize:
+                nmags = (smags - npmedian(smags))/npstd(smags)
+            else:
+                nmags = smags
+
+            tasks = [(stimes, nmags, serrs, x, phasebinsize, mindetperbin)
+                     for x in frequencies]
+
+            lsp = pool.map(stellingwerf_pdm_worker, tasks)
+
+            pool.close()
+            pool.join()
+            del pool
+
+            lsp = nparray(lsp)
+            periods = 1.0/frequencies
+
+            # find the nbestpeaks for the periodogram: 1. sort the lsp array by
+            # lowest value first 2. go down the values until we find five values
+            # that are separated by at least periodepsilon in period
+
+            # make sure to filter out the non-finite values of lsp
+            finitepeakind = npisfinite(lsp)
+            finlsp = lsp[finitepeakind]
+            finperiods = periods[finitepeakind]
+
+
+            bestperiodind = npargmin(finlsp)
+
+            sortedlspind = np.argsort(finlsp)
+            sortedlspperiods = finperiods[sortedlspind]
+            sortedlspvals = finlsp[sortedlspind]
+
+            prevbestlspval = sortedlspvals[0]
+            # now get the nbestpeaks
+            nbestperiods, nbestlspvals, peakcount = (
+                [finperiods[bestperiodind]],
+                [finlsp[bestperiodind]],
+                1
+            )
+            prevperiod = sortedlspperiods[0]
+
+            # find the best nbestpeaks in the lsp and their periods
+            for period, lspval in zip(sortedlspperiods, sortedlspvals):
+
+                if peakcount == nbestpeaks:
+                    break
+                perioddiff = abs(period - prevperiod)
+                bestperiodsdiff = [abs(period - x) for x in nbestperiods]
+
+                # print('prevperiod = %s, thisperiod = %s, '
+                #       'perioddiff = %s, peakcount = %s' %
+                #       (prevperiod, period, perioddiff, peakcount))
+
+                # this ensures that this period is different from the last
+                # period and from all the other existing best periods by
+                # periodepsilon to make sure we jump to an entire different peak
+                # in the periodogram
+                if (perioddiff > periodepsilon and
+                    all(x > periodepsilon for x in bestperiodsdiff)):
+                    nbestperiods.append(period)
+                    nbestlspvals.append(lspval)
+                    peakcount = peakcount + 1
+
+                prevperiod = period
+
+
+            return {'bestperiod':finperiods[bestperiodind],
+                    'bestlspval':finlsp[bestperiodind],
+                    'nbestpeaks':nbestpeaks,
+                    'nbestlspvals':nbestlspvals,
+                    'nbestperiods':nbestperiods,
+                    'lspvals':lsp,
+                    'periods':periods,
+                    'method':'pdm'}
+
+        else:
+
+            LOGERROR('no good detections for these times and mags, skipping...')
+            return {'bestperiod':npnan,
+                    'bestlspval':npnan,
+                    'nbestpeaks':nbestpeaks,
+                    'nbestlspvals':None,
+                    'nbestperiods':None,
+                    'lspvals':None,
+                    'periods':None,
+                    'method':'pdm'}
+    else:
+
+        LOGERROR('no good detections for these times and mags, skipping...')
+        return {'bestperiod':npnan,
+                'bestlspval':npnan,
+                'nbestpeaks':nbestpeaks,
+                'nbestlspvals':None,
+                'nbestperiods':None,
+                'lspvals':None,
+                'periods':None,
+                'method':'pdm'}
