+++ astrobase/periodbase_to_be_removed.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+
+'''
+periodbase.py - Waqas Bhatti (wbhatti@astro.princeton.edu) - Dec 2015
+
+Contains various useful tools for period finding.
+
+
+'''
